Correcting help text pointing to Glossary
I see some additional text was added pointing people to the glossary through the menu system.

So I tried it, and it's not there. It is under "Get Started", not "Core Concepts" although I guess if you're guessing that's where it should be then it probably should be, I mean how many people "Get Started" by reading through a Glossary? "Hey Steve, watchya doing over there?" "Oh, I'm learning Japanese, I thought I'd start by learning ALL THE WORDS". 

Anyway, might be an idea to actually check before committing, right?

I do like this though, I, like you, didn't know where it was in the menu system and, like you, didn't bother to go find it either so it is good to let the reader know how to access it. Providing of course you tell them the right place, and of course if the right place is in a logical place then that would help too.

Have a nice day ;)

sbp